Backend/requirements.txt
agency_swarm==0.2.5
gradio==4.37.2
openai==1.30.4
pydantic==2.8.0
pydifact==0.1.7
python-dotenv==1.0.1
Requests==2.32.3
typing_extensions==4.12.2
uvicorn==0.30.1
